Backend/requirements.txt
supabase
fastapi
uvicorn[standard]
python-multipart
openai
openai-whisper
python-dotenv
ffmpeg-python